Move utils,types and ApiConstants
