+++ src/tests/api/ApiUtils.ts
@@ -1,28 +1,28 @@
 import Ajv from "ajv";
 import { XMLParser } from "fast-xml-parser";
-import { HARNESS_API_INSTANCE } from "./ApiTestSteps";
-import { TxmaEvent, TxmaEventName } from "../../../utils/TxmaEvent";
-import * as F2F_CRI_AUTH_CODE_ISSUED_SCHEMA from "../../data/F2F_CRI_AUTH_CODE_ISSUED_SCHEMA.json";
-import * as F2F_CRI_END_SCHEMA from "../../data/F2F_CRI_END_SCHEMA.json";
-import * as F2F_CRI_START_SCHEMA from "../../data/F2F_CRI_START_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_00_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_00_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_01_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_01_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_02_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_02_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_03_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_03_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_04_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_04_SCHEMA.json";
-import * as F2F_CRI_VC_ISSUED_05_SCHEMA from "../../data/F2F_CRI_VC_ISSUED_05_SCHEMA.json";
-import * as F2F_DOCUMENT_UPLOADED_SCHEMA from "../../data/F2F_DOCUMENT_UPLOADED_SCHEMA.json";
-import * as F2F_YOTI_PDF_EMAILED_SCHEMA from "../../data/F2F_YOTI_PDF_EMAILED_SCHEMA.json";
-import * as F2F_YOTI_START_SCHEMA from "../../data/F2F_YOTI_START_SCHEMA.json";
-import * as F2F_YOTI_RESPONSE_RECEIVED_SCHEMA from "../../data/F2F_YOTI_RESPONSE_RECEIVED_SCHEMA.json";
-import * as F2F_YOTI_START_00_SCHEMA from "../../data/F2F_YOTI_START_00_SCHEMA.json";
-import * as F2F_YOTI_START_01_SCHEMA from "../../data/F2F_YOTI_START_01_SCHEMA.json";
-import * as F2F_YOTI_START_02_SCHEMA from "../../data/F2F_YOTI_START_02_SCHEMA.json";
-import * as F2F_YOTI_START_03_SCHEMA from "../../data/F2F_YOTI_START_03_SCHEMA.json";
-import * as F2F_YOTI_START_04_SCHEMA from "../../data/F2F_YOTI_START_04_SCHEMA.json";
-import * as F2F_YOTI_START_05_SCHEMA from "../../data/F2F_YOTI_START_05_SCHEMA.json";
-import * as F2F_CRI_SESSION_ABORTED_SCHEMA from "../../data/F2F_CRI_SESSION_ABORTED_SCHEMA.json";
+import { HARNESS_API_INSTANCE } from "./backend/ApiTestSteps";
+import { TxmaEvent, TxmaEventName } from "../../utils/TxmaEvent";
+import * as F2F_CRI_AUTH_CODE_ISSUED_SCHEMA from "../data/F2F_CRI_AUTH_CODE_ISSUED_SCHEMA.json";
+import * as F2F_CRI_END_SCHEMA from "../data/F2F_CRI_END_SCHEMA.json";
+import * as F2F_CRI_START_SCHEMA from "../data/F2F_CRI_START_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_SCHEMA from "../data/F2F_CRI_VC_ISSUED_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_00_SCHEMA from "../data/F2F_CRI_VC_ISSUED_00_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_01_SCHEMA from "../data/F2F_CRI_VC_ISSUED_01_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_02_SCHEMA from "../data/F2F_CRI_VC_ISSUED_02_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_03_SCHEMA from "../data/F2F_CRI_VC_ISSUED_03_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_04_SCHEMA from "../data/F2F_CRI_VC_ISSUED_04_SCHEMA.json";
+import * as F2F_CRI_VC_ISSUED_05_SCHEMA from "../data/F2F_CRI_VC_ISSUED_05_SCHEMA.json";
+import * as F2F_DOCUMENT_UPLOADED_SCHEMA from "../data/F2F_DOCUMENT_UPLOADED_SCHEMA.json";
+import * as F2F_YOTI_PDF_EMAILED_SCHEMA from "../data/F2F_YOTI_PDF_EMAILED_SCHEMA.json";
+import * as F2F_YOTI_START_SCHEMA from "../data/F2F_YOTI_START_SCHEMA.json";
+import * as F2F_YOTI_RESPONSE_RECEIVED_SCHEMA from "../data/F2F_YOTI_RESPONSE_RECEIVED_SCHEMA.json";
+import * as F2F_YOTI_START_00_SCHEMA from "../data/F2F_YOTI_START_00_SCHEMA.json";
+import * as F2F_YOTI_START_01_SCHEMA from "../data/F2F_YOTI_START_01_SCHEMA.json";
+import * as F2F_YOTI_START_02_SCHEMA from "../data/F2F_YOTI_START_02_SCHEMA.json";
+import * as F2F_YOTI_START_03_SCHEMA from "../data/F2F_YOTI_START_03_SCHEMA.json";
+import * as F2F_YOTI_START_04_SCHEMA from "../data/F2F_YOTI_START_04_SCHEMA.json";
+import * as F2F_YOTI_START_05_SCHEMA from "../data/F2F_YOTI_START_05_SCHEMA.json";
+import * as F2F_CRI_SESSION_ABORTED_SCHEMA from "../data/F2F_CRI_SESSION_ABORTED_SCHEMA.json";
 
 const ajv = new Ajv({ strictTuples: false });
 ajv.addSchema(F2F_CRI_AUTH_CODE_ISSUED_SCHEMA, "F2F_CRI_AUTH_CODE_ISSUED_SCHEMA");
