+++ src/tests/api/backend/UnhappyPath.test.ts
@@ -17,7 +17,7 @@ import dataPassportMissingFad from "../../data/dataPassportMissingFad.json";
 import dataPassportBlankFad from "../../data/dataPassportBlankFad.json";
 import dataPassportIncorrectFad from "../../data/dataPassportIncorrectFad.json";
 import dataPassportInvalidFadFormat from "../../data/dataPassportInvalidFadFormat.json";
-import { StubStartRequest, DocSelectionData } from "./types";
+import { StubStartRequest, DocSelectionData } from "../types";
 import { randomUUID } from "crypto";
